requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
